refactory: Alram 컴포넌트 간소화
